Move nzbget base entity to separate module
